requirements.txt
paperswithcode-client==0.3.1
Flask==2.3.2